projects/company_handbook_assistant/requirements.txt
gitpython
superpowered-sdk